+++ PersistentStorage/Sources/Data/Repository/PersistentStorage/MainPersistentStorage.swift
@@ -53,10 +53,7 @@ public class MainPersistentStorage: PersistentStorage {
     public func remove(
         valueKey: String
     ) -> AnyPublisher<Void, PersistentStorageError> {
-        userDefaults.set(
-            nil,
-            forKey: valueKey
-        )
+        userDefaults.removeObject(forKey: valueKey)
         log(
             message: "✅ Successfully removed value with key {\(valueKey)}",
             for: .debug
@@ -65,25 +62,6 @@ public class MainPersistentStorage: PersistentStorage {
         return Just(())
             .setFailureType(to: PersistentStorageError.self)
             .eraseToAnyPublisher()
-    }
-
-    /// Method for reading a value with given key and type.
-    ///
-    /// - Possible failures:
-    ///     - If type resolution with given parameter `valueType` fails, method will return `noValueFoundWithGivenType` failure.
-    ///     - If no value is associated with key given in parameter `valueKey` or its value is nil, the method will return `noValueFound` failure.
-    ///
-    /// - Parameters:
-    ///     - valueType: The type of value that you wish to read.
-    ///     - valueKey: Key that is assigned to a value that you wish to read.
-    public func read<T>(
-        valueType: T.Type,
-        valueKey: String
-    ) throws -> T {
-        try getPersistedValue(
-            valueType: valueType,
-            valueKey: valueKey
-        )
     }
 
     /// Method for reading a value with given key and type. Result is returned as a publisher.